+++ packages/dust_db_sqlite3/lib/src/sqlite_pool.dart
@@ -1,3 +1,6 @@
+import 'dart:convert';
+import 'dart:typed_data';
+
 import 'package:dust_dart/db.dart';
 import 'package:sqlite3/sqlite3.dart' as sqlite;